+++ frontend/src/metabase/parameters/components/ParameterValuePicker/ListPicker/index.ts
@@ -1,0 +1,1 @@
+export { ListPicker } from "./ListPicker";
