+++ frontend/src/metabase/parameters/components/ParameterValuePicker/ListPickerConnected/index.ts
@@ -1,0 +1,1 @@
+export { ListPickerConnected } from "./ListPickerConnected";
