+++ frontend/src/metabase/parameters/components/ParameterValuePicker/OwnDatePicker/OwnDatePicker.tsx
@@ -13,7 +13,7 @@ import type {
   ParameterType,
 } from "metabase-types/api";
 
-import { PickerIcon, TextInputTrirgger } from "./ParameterValuePicker.styled";
+import { PickerIcon, TextInputTrirgger } from "../ParameterValuePicker.styled";
 
 interface OwnDatePickerProps {
   value: string;
