+++ frontend/src/metabase/parameters/components/ParameterValuePicker/OwnDatePicker/index.ts
@@ -1,0 +1,1 @@
+export { OwnDatePicker } from "./OwnDatePicker";
