+++ frontend/src/metabase/parameters/components/ParameterValuePicker/PlainValueInput/index.ts
@@ -1,0 +1,1 @@
+export { PlainValueInput } from "./PlainValueInput";
