+++ frontend/src/metabase/parameters/components/ParameterValuePicker/util.ts
